auth-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
google-auth>=2.22.0
httpx>=0.24.0
requests>=2.31.0
hypothesis>=6.82.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
